Refactored annotation api route to controller
